+++ core/core-api/src/functionalTest/groovy/com/hbc/pms/core/api/services/MonitoringFunctionalSpec.groovy
@@ -2,7 +2,7 @@ package com.hbc.pms.core.api.services
 
 import com.hbc.pms.core.api.FunctionalTestSpec
 import com.hbc.pms.core.api.TestDataFixture
-import com.hbc.pms.core.api.service.SensorConfigurationPersistenceService
+import com.hbc.pms.core.api.service.blueprint.SensorConfigurationPersistenceService
 import com.hbc.pms.core.model.SensorConfiguration
 import com.hbc.pms.integration.db.repository.BlueprintRepository
 import com.hbc.pms.integration.db.repository.SensorConfigurationRepository
