+++ core/core-api/src/functionalTest/groovy/com/hbc/pms/core/api/services/NotificationServiceIntegrationSpec.groovy
@@ -3,10 +3,10 @@ package com.hbc.pms.core.api.services
 import com.hbc.pms.core.api.FunctionalTestSpec
 import com.hbc.pms.core.api.TestDataFixture
 import com.hbc.pms.core.api.controller.v1.WebSocketController
-import com.hbc.pms.core.api.service.AlarmConditionService
-import com.hbc.pms.core.api.service.AlarmPersistenceService
 import com.hbc.pms.core.api.service.WebSocketService
-import com.hbc.pms.core.api.support.notification.PopupChannel
+import com.hbc.pms.core.api.service.alarm.AlarmConditionService
+import com.hbc.pms.core.api.service.alarm.AlarmPersistenceService
+import com.hbc.pms.core.api.service.alarm.notification.PopupChannel
 import com.hbc.pms.core.model.enums.AlarmStatus
 import com.hbc.pms.core.model.enums.AlarmType
 import com.hbc.pms.integration.db.repository.*
