+++ core/core-api/src/main/java/com/hbc/pms/core/api/config/plc/PlcMonitoringDataSource.java
@@ -1,6 +1,6 @@
 package com.hbc.pms.core.api.config.plc;
 
-import com.hbc.pms.core.api.service.BlueprintPersistenceService;
+import com.hbc.pms.core.api.service.blueprint.BlueprintPersistenceService;
 import com.hbc.pms.core.model.enums.BlueprintType;
 import com.hbc.pms.plc.api.scraper.HbcScrapeJobDataSource;
 import java.util.List;
