+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/AlarmConditionController.java
@@ -4,10 +4,10 @@ import com.hbc.pms.core.api.controller.v1.request.CreateAlarmConditionCommand;
 import com.hbc.pms.core.api.controller.v1.request.UpdateActionMessagesCommand;
 import com.hbc.pms.core.api.controller.v1.request.UpdateAlarmConditionCommand;
 import com.hbc.pms.core.api.controller.v1.response.AlarmConditionResponse;
-import com.hbc.pms.core.api.service.AlarmActionPersistenceService;
-import com.hbc.pms.core.api.service.AlarmActionService;
-import com.hbc.pms.core.api.service.AlarmConditionPersistenceService;
-import com.hbc.pms.core.api.service.AlarmConditionService;
+import com.hbc.pms.core.api.service.alarm.AlarmActionPersistenceService;
+import com.hbc.pms.core.api.service.alarm.AlarmActionService;
+import com.hbc.pms.core.api.service.alarm.AlarmConditionPersistenceService;
+import com.hbc.pms.core.api.service.alarm.AlarmConditionService;
 import com.hbc.pms.core.model.AlarmAction;
 import com.hbc.pms.core.model.AlarmCondition;
 import com.hbc.pms.support.web.response.ApiResponse;
