+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/AlarmHistoryController.java
@@ -1,7 +1,7 @@
 package com.hbc.pms.core.api.controller.v1;
 
 import com.hbc.pms.core.api.controller.v1.response.AlarmHistoryResponse;
-import com.hbc.pms.core.api.service.AlarmPersistenceService;
+import com.hbc.pms.core.api.service.alarm.AlarmPersistenceService;
 import com.hbc.pms.core.model.enums.AlarmStatus;
 import com.hbc.pms.support.web.response.ApiResponse;
 import java.util.List;
