+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/BlueprintController.java
@@ -7,8 +7,8 @@ import com.hbc.pms.core.api.controller.v1.request.SearchBlueprintCommand;
 import com.hbc.pms.core.api.controller.v1.request.SensorConfigurationRequest;
 import com.hbc.pms.core.api.controller.v1.request.UpdateSensorConfigurationCommand;
 import com.hbc.pms.core.api.controller.v1.response.BlueprintResponse;
-import com.hbc.pms.core.api.service.BlueprintPersistenceService;
-import com.hbc.pms.core.api.service.SensorConfigurationPersistenceService;
+import com.hbc.pms.core.api.service.blueprint.BlueprintPersistenceService;
+import com.hbc.pms.core.api.service.blueprint.SensorConfigurationPersistenceService;
 import com.hbc.pms.core.model.Blueprint;
 import com.hbc.pms.core.model.SensorConfiguration;
 import com.hbc.pms.support.web.response.ApiResponse;
