+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/ReportController.java
@@ -1,15 +1,17 @@
 package com.hbc.pms.core.api.controller.v1;
 
-import com.hbc.pms.core.api.controller.v1.common.Page;
 import com.hbc.pms.core.api.controller.v1.response.ReportResponse;
-import com.hbc.pms.core.api.service.ReportPersistenceService;
-import com.hbc.pms.core.api.service.ReportTypePersistenceService;
+import com.hbc.pms.core.api.service.report.ReportPersistenceService;
+import com.hbc.pms.core.api.service.report.ReportTypePersistenceService;
+import com.hbc.pms.core.model.Report;
 import com.hbc.pms.core.model.ReportType;
 import com.hbc.pms.core.model.criteria.ReportCriteria;
+import com.hbc.pms.support.web.pagination.QueryResult;
 import com.hbc.pms.support.web.response.ApiResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -31,20 +33,14 @@ public class ReportController {
   }
 
   @GetMapping("details")
-  public ApiResponse<Page<ReportResponse>> getDetails(
+  public ApiResponse<QueryResult<ReportResponse>> getDetails(
       @RequestParam(value = "page", defaultValue = "1") int page,
       @RequestParam(value = "size", defaultValue = "10") int size,
       ReportCriteria filter) {
-    var reports = reportPersistenceService.getAll(filter, PageRequest.of(page - 1, size));
+    Page<Report> reports = reportPersistenceService.getAll(filter, PageRequest.of(page - 1, size));
+
     return ApiResponse.success(
-        Page.<ReportResponse>builder()
-            .pageTotal(reports.getPageTotal())
-            .recordTotal(reports.getRecordTotal())
-            .content(
-                reports.getContent().stream()
-                    .map(el -> mapper.map(el, ReportResponse.class))
-                    .toList())
-            .build());
+        QueryResult.fromPage(reports.map(report -> mapper.map(report, ReportResponse.class))));
   }
 
   @GetMapping("details/{id}")
