+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/SensorConfigurationController.java
@@ -2,7 +2,7 @@ package com.hbc.pms.core.api.controller.v1;
 
 import com.hbc.pms.core.api.controller.v1.request.SearchBlueprintCommand;
 import com.hbc.pms.core.api.controller.v1.response.SensorConfigurationResponse;
-import com.hbc.pms.core.api.service.SensorConfigurationPersistenceService;
+import com.hbc.pms.core.api.service.blueprint.SensorConfigurationPersistenceService;
 import com.hbc.pms.support.web.response.ApiResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
