+++ core/core-api/src/main/java/com/hbc/pms/core/api/controller/v1/request/auth/QueryUserCommand.java
@@ -1,0 +1,5 @@
+package com.hbc.pms.core.api.controller.v1.request.auth;
+
+import com.hbc.pms.support.web.pagination.PaginationCommand;
+
+public class QueryUserCommand extends PaginationCommand {}
