+++ core/core-api/src/main/java/com/hbc/pms/core/api/service/AbstractPersistenceService.java
@@ -1,0 +1,28 @@
+package com.hbc.pms.core.api.service;
+
+import java.util.Collection;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+
+public abstract class AbstractPersistenceService<V> {
+  @Autowired
+  protected ModelMapper mapper;
+
+  public <K> K mapToModel(V entity, Class<K> clazzModel) {
+    return mapper.map(entity, clazzModel);
+  }
+  public <K> Collection<K> mapToModel(Collection<V> entities, Class<K> clazzModel) {
+    return entities.stream().map(entity -> mapToModel(entity, clazzModel)).toList();
+  }
+  public <K> Page<K> mapToModel(Page<V> entities, Class<K> clazzModel) {
+    return entities.map(entity -> mapToModel(entity, clazzModel));
+  }
+  public <K> V mapToEntity(K model, Class<V> clazzEntity) {
+    return mapper.map(model, clazzEntity);
+  }
+  public <K> Collection<V> mapToEntity(Collection<K> models, Class<V> clazzEntity) {
+    return models.stream().map(model -> mapToEntity(model, clazzEntity)).toList();
+  }
+
+}
