+++ core/core-api/src/main/java/com/hbc/pms/core/api/service/alarm/AlarmActionService.java
@@ -1,4 +1,4 @@
-package com.hbc.pms.core.api.service;
+package com.hbc.pms.core.api.service.alarm;
 
 import com.hbc.pms.core.model.AlarmAction;
 import com.hbc.pms.core.model.AlarmCondition;
