+++ core/core-api/src/main/java/com/hbc/pms/core/api/service/auth/UserService.java
@@ -1,8 +1,12 @@
 package com.hbc.pms.core.api.service.auth;
 
+import com.hbc.pms.core.api.controller.v1.request.auth.QueryUserCommand;
 import com.hbc.pms.core.model.User;
+import com.hbc.pms.support.web.pagination.QueryResult;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -18,6 +22,12 @@ public class UserService implements UserDetailsService {
   @Override
   public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
     return userPersistenceService.findByUsername(username);
+  }
+
+  public QueryResult<User> query(QueryUserCommand queryUserCommand) {
+    Pageable pageable = queryUserCommand.toPageable();
+    Page<User> queryResult = userPersistenceService.query(pageable);
+    return QueryResult.fromPage(queryResult);
   }
 
   public User findById(Long userId) {
