+++ support/web/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
   implementation "org.springframework.boot:spring-boot-starter-web"
+  implementation "org.springframework.data:spring-data-commons"
 }
